requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0

# CLIP and open_clip
open-clip-torch>=2.20.0

# Machine learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bar
tqdm>=4.65.0

# Data processing
opencv-python>=4.8.0

# Optional: for better performance
# Install with: pip install -r requirements.txt
